+++ web_music_player_project-master/src/main/java/com/example/servicio/CancionService.java
@@ -22,7 +22,7 @@ public class CancionService {
         }
     }
 
-    public Song get(Long id) {
+    public Song get(Integer id) {
         return songRepository.find(id);
     }
